v1.00 rev3921 (7 Sept 2019)
- Unsupported EA videos caused error
Known issues:
- Videos with variable frame rates cause a lot of problems:
  * Usually unable to detect when videos have a variable frame rate
  * The audio and video become out of sync when saving a video with
	variable frame rates as AVI
  * Even when a video is identified as having a variable frame rate,
	some video players do not play the generated AVI correctly (primarily
	on Windows). Try a more reliable player (such as VLC media player).
- Real-time playback doesn't work for videos with a long audio delay
- Real-time playback stops early if audio ends before video ends

+++ jpsxdec/src/jpsxdec/psxvideo/bitstreams/BitStreamUncompressor_Iki.java
@@ -606,7 +606,7 @@ public class BitStreamUncompressor_Iki extends BitStreamUncompressor {
             ab = _bottom8.toByteArray();
 
             byte[] abHdr = new byte[10];
-            IO.writeInt16LE(abHdr, 0, (short)Calc.calculateHalfCeiling32(iMdecCodeCount));
+            IO.writeInt16LE(abHdr, 0, Calc.calculateHalfCeiling32(iMdecCodeCount));
             IO.writeInt16LE(abHdr, 2, (short)0x3800);
             IO.writeInt16LE(abHdr, 4, (short)_iWidth);
             IO.writeInt16LE(abHdr, 6, (short)_iHeight);
